pyproject.toml
[build-system]
requires = ["flit_core>=3.4,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "neuronic"
authors = [{name = "Nidal Alhariri", email = "level09@gmail.com"}]
dynamic = ["version", "description"]
readme = "README.md"
license = {file = "LICENSE"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
keywords = ["data", "transformation", "AI", "OpenAI", "GPT"]
requires-python = ">=3.8"
dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=0.19.0"
]

[project.urls]
Home = "https://github.com/level09/neuronic"
Documentation = "https://github.com/level09/neuronic#readme"
Source = "https://github.com/level09/neuronic"

[project.optional-dependencies]
dev = [
    "black>=22.0",
    "isort>=5.0",
    "flake8>=4.0",
] 